backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
chromadb
sentence-transformers
torch
transformers
openai
pypdf
python-docx
tiktoken
pandas
numpy
aiosqlite
tqdm
